pyproject.toml
[project]
name = "replicate-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "[Jamie Borland]",email = "[s2059547]@sms.ed.ac.uk"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py (>=2.4.0,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "replicate (>=1.0.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
